# api/openapi.yaml
openapi: 3.0.3
info:
  title: EduCloud Lite API
  version: 0.1.0
  description: Skeleton API contract for EduCloud Lite.
servers:
  - url: http://localhost:8000/api
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      responses:
        "200":
          description: User registered
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      responses:
        "200":
          description: User logged in
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
  /courses:
    get:
      tags: [Courses]
      summary: Get course list
      responses:
        "200":
          description: Course list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Course"
    post:
      tags: [Courses]
      summary: Create course (instructor role required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseCreate"
      responses:
        "200":
          description: Course created
        "403":
          description: Instructor role required
  /courses/{course_id}:
    get:
      tags: [Courses]
      summary: Get course detail with lessons
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Course not found
    put:
      tags: [Courses]
      summary: Update course (owner instructor or admin only)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Course updated
        "403":
          description: Not the course owner or admin
        "404":
          description: Course not found
    delete:
      tags: [Courses]
      summary: Delete course (owner instructor or admin only)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course deleted
        "403":
          description: Not the course owner or admin
        "404":
          description: Course not found
  /courses/{course_id}/lessons:
    get:
      tags: [Lessons]
      summary: Get lessons by course, ordered by order_index
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Lesson list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Lesson"
        "404":
          description: Course not found
    post:
      tags: [Lessons]
      summary: Create lesson (course owner instructor or admin only)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonCreate"
      responses:
        "200":
          description: Lesson created
        "403":
          description: Not the course owner or admin
        "404":
          description: Course not found
  /lessons/{lesson_id}:
    put:
      tags: [Lessons]
      summary: Update lesson (course owner instructor or admin only)
      parameters:
        - $ref: "#/components/parameters/LessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Lesson updated
        "403":
          description: Not the course owner or admin
        "404":
          description: Lesson not found
    delete:
      tags: [Lessons]
      summary: Delete lesson (course owner instructor or admin only)
      parameters:
        - $ref: "#/components/parameters/LessonId"
      responses:
        "200":
          description: Lesson deleted
        "403":
          description: Not the course owner or admin
        "404":
          description: Lesson not found
  /courses/{course_id}/enroll:
    post:
      tags: [Enrollment]
      summary: Enroll course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course enrolled
  /my-courses:
    get:
      tags: [Enrollment]
      summary: Get enrolled courses
      responses:
        "200":
          description: Enrolled courses
  /lessons/{lesson_id}/complete:
    post:
      tags: [Progress]
      summary: Complete lesson
      parameters:
        - $ref: "#/components/parameters/LessonId"
      responses:
        "200":
          description: Lesson completed
  /courses/{course_id}/progress:
    get:
      tags: [Progress]
      summary: Get course progress
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course progress
  /upload/course-thumbnail:
    post:
      tags: [Upload]
      summary: Upload course thumbnail
      responses:
        "200":
          description: Thumbnail uploaded
  /upload/lesson-material:
    post:
      tags: [Upload]
      summary: Upload lesson material
      responses:
        "200":
          description: Material uploaded
  /upload/video:
    post:
      tags: [Upload]
      summary: Upload video
      responses:
        "200":
          description: Video uploaded
components:
  parameters:
    CourseId:
      name: course_id
      in: path
      required: true
      schema:
        type: integer
    LessonId:
      name: lesson_id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
    CourseCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        level:
          type: string
          default: All levels
        category:
          type: string
          default: EduCloud
        learning_outcomes:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        thumbnail_url:
          type: string
          nullable: true
        price:
          type: number
          default: 0
        status:
          type: string
          enum: [draft, published, hidden, archived]
          default: draft
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        level:
          type: string
        category:
          type: string
        learning_outcomes:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        thumbnail_url:
          type: string
          nullable: true
        price:
          type: number
        status:
          type: string
          enum: [draft, published, hidden, archived]
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        level:
          type: string
        category:
          type: string
        learning_outcomes:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        thumbnail_url:
          type: string
          nullable: true
        price:
          type: number
        status:
          type: string
          enum: [draft, published, hidden, archived]
        instructor_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
    LessonCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
          nullable: true
        video_url:
          type: string
          nullable: true
        material_url:
          type: string
          nullable: true
        order_index:
          type: integer
          default: 0
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
          nullable: true
        video_url:
          type: string
          nullable: true
        material_url:
          type: string
          nullable: true
        order_index:
          type: integer
    Lesson:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        title:
          type: string
        content:
          type: string
          nullable: true
        video_url:
          type: string
          nullable: true
        material_url:
          type: string
          nullable: true
        order_index:
          type: integer